Database Authentication
Authentication is no longer done by Memory Authentication, the Users information is now securely stored in a Database and can be accessed to provide authentication. Each 'User' stored in the database has a Many to Many Relationship with 'Roles' joined by Users_Roles table. This is then accessed in the application which allows for the method level authorisation to work based on the user signed in/accessing the api, though at this moment, only passed via Postman.